requirements.txt
streamlit
llama-index
llama-parse
python-dotenv
openai
pydantic
nest_asyncio
